Change webpack config
